drfElevator/requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.3.0
click==8.1.3
Django==4.2
djangorestframework==3.14.0
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
